fix: Made it possible to use signed integers for creating PlcTIME values in Go (Needed to add explicit casts to the generated Go gode)

+++ plc4go/protocols/abeth/readwrite/model/DF1RequestMessage.go
@@ -267,14 +267,14 @@ func (pm *_DF1RequestMessage) SerializeParent(ctx context.Context, writeBuffer u
 
 	// Simple Field (destinationAddress)
 	destinationAddress := uint8(m.GetDestinationAddress())
-	_destinationAddressErr := writeBuffer.WriteUint8("destinationAddress", 8, (destinationAddress))
+	_destinationAddressErr := writeBuffer.WriteUint8("destinationAddress", 8, uint8((destinationAddress)))
 	if _destinationAddressErr != nil {
 		return errors.Wrap(_destinationAddressErr, "Error serializing 'destinationAddress' field")
 	}
 
 	// Simple Field (sourceAddress)
 	sourceAddress := uint8(m.GetSourceAddress())
-	_sourceAddressErr := writeBuffer.WriteUint8("sourceAddress", 8, (sourceAddress))
+	_sourceAddressErr := writeBuffer.WriteUint8("sourceAddress", 8, uint8((sourceAddress)))
 	if _sourceAddressErr != nil {
 		return errors.Wrap(_sourceAddressErr, "Error serializing 'sourceAddress' field")
 	}
@@ -289,7 +289,7 @@ func (pm *_DF1RequestMessage) SerializeParent(ctx context.Context, writeBuffer u
 			}).Msg("Overriding reserved field with unexpected value.")
 			reserved = *pm.reservedField0
 		}
-		_err := writeBuffer.WriteUint16("reserved", 16, reserved)
+		_err := writeBuffer.WriteUint16("reserved", 16, uint16(reserved))
 		if _err != nil {
 			return errors.Wrap(_err, "Error serializing 'reserved' field")
 		}
@@ -297,7 +297,7 @@ func (pm *_DF1RequestMessage) SerializeParent(ctx context.Context, writeBuffer u
 
 	// Discriminator Field (commandCode) (Used as input to a switch field)
 	commandCode := uint8(child.GetCommandCode())
-	_commandCodeErr := writeBuffer.WriteUint8("commandCode", 8, (commandCode))
+	_commandCodeErr := writeBuffer.WriteUint8("commandCode", 8, uint8((commandCode)))
 
 	if _commandCodeErr != nil {
 		return errors.Wrap(_commandCodeErr, "Error serializing 'commandCode' field")
@@ -305,14 +305,14 @@ func (pm *_DF1RequestMessage) SerializeParent(ctx context.Context, writeBuffer u
 
 	// Simple Field (status)
 	status := uint8(m.GetStatus())
-	_statusErr := writeBuffer.WriteUint8("status", 8, (status))
+	_statusErr := writeBuffer.WriteUint8("status", 8, uint8((status)))
 	if _statusErr != nil {
 		return errors.Wrap(_statusErr, "Error serializing 'status' field")
 	}
 
 	// Simple Field (transactionCounter)
 	transactionCounter := uint16(m.GetTransactionCounter())
-	_transactionCounterErr := writeBuffer.WriteUint16("transactionCounter", 16, (transactionCounter))
+	_transactionCounterErr := writeBuffer.WriteUint16("transactionCounter", 16, uint16((transactionCounter)))
 	if _transactionCounterErr != nil {
 		return errors.Wrap(_transactionCounterErr, "Error serializing 'transactionCounter' field")
 	}
